% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_accession_contigs_megahit_dir.R
\name{add_accession_contigs_megahit_dir}
\alias{add_accession_contigs_megahit_dir}
\title{Function to concatenate megahit final contigs into one fasta file with sample accession added to the headers.
This function can be used to create one file for input in makeblastdb to create a blast database of all the final contigs.}
\usage{
add_accession_contigs_megahit_dir(megahit_dir, outfile)
}
\arguments{
\item{megahit_dir}{String, path to directorty in which the fastq files are present, here the blast and contigs folder will be made.}

\item{outfile}{String, path to output file to create.}
}
\value{
This functions creates the output file specified, containing the sequencing in fasta format.
}
\description{
Function to concatenate megahit final contigs into one fasta file with sample accession added to the headers.
This function can be used to create one file for input in makeblastdb to create a blast database of all the final contigs.
}
\examples{
\dontrun{
add_accession_contigs_megahit(megahit_dir = "/home/$USER/research_drive/geodescent/samples/MGYS00000974/megahit",
                              outfile = "/home/$USER/research_drive/geodescent/samples/MGYS00000974/blast/contigs/all_samples2.fa")
                              }
}
