% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast}
\alias{blast}
\title{This function is a wrapper for the blast+ suite tools.}
\usage{
blast(blast, blast_db, input, out, e = 0.01, format = 6)
}
\arguments{
\item{blast}{String, type of blast search to perform, e.g. "tblastn" or "blastp"}

\item{blast_db}{String, path to blast database (you can create a database useing the makeblastdb function)}

\item{input}{String, path to input query file (sequence to blast against the database)}

\item{e}{Integer, evalue cutoff to apply, default = 0.01}

\item{format}{Integer, format type to use for results, default = 6 (tabular). Options:
alignment view options:
  0 = pairwise,
  1 = query-anchored showing identities,
  2 = query-anchored no identities,
  3 = flat query-anchored, show identities,
  4 = flat query-anchored, no identities,
  5 = XML Blast output,
  6 = tabular,
  7 = tabular with comment lines,
  8 = Text ASN.1,
  9 = Binary ASN.1,
  10 = Comma-separated values,
  11 = BLAST archive format (ASN.1)}
}
\value{
blast_out File with blast results
}
\description{
This function is a wrapper for the blast+ suite tools.
}
\examples{
\dontrun{
blast(blast = "tblastn",
blast_db = "/home/rstudio/data/tara_db/tara_db",
input = "/home/rstudio/scpackage/inst/extdata/sc_markers.faa",
out = "/home/rstudio/data/test_tara_10n2",
format = 6)
}
}
