% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_rename_seq-id.R
\name{cat_rename_seq_id}
\alias{cat_rename_seq_id}
\title{Concatenate binned contigs into one file per sample and add samplename and bin name to the headers.}
\usage{
cat_rename_seq_id(outdir, sample_accession)
}
\arguments{
\item{bin_dirs_parent_dir}{String, path to directory containing directories with bins (metabat2 folder in pipeline function output)}
}
\value{
A file with the concatenated sequences with renamed headers is created in each sample folder. (total_(samplename).fa)
}
\description{
Concatenate binned contigs into one file per sample and add samplename and bin name to the headers.
}
\examples{
\dontrun{
metabat_dir = paste0(samples_dir, "/metabat2")
cat_rename_seq_id(bin_dirs_parent_dir = metabat_dir)
}
}
