% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_multisummary.R
\name{fastqc_multisummary}
\alias{fastqc_multisummary}
\title{Plots sum stats of multiple fastqc results}
\usage{
fastqc_multisummary(results_dir)
}
\arguments{
\item{results_dir}{Path to directory containing fastqc results}
}
\value{

}
\description{
Plots sum stats of multiple fastqc results
}
\examples{
fastqc_multisummary(results_dir = paste0(samples_dir, "/fastqc/untrimmed"))
}
