% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbtools.R
\name{gbtools_plot}
\alias{gbtools_plot}
\title{Gbtools bin plot for Shiny}
\usage{
gbtools_plot(covstats_datapath, taxonomy_datapath, colour_tax_level = "Genus")
}
\arguments{
\item{colour_tax_level}{Taxonomy level to colour contigs with, e.g. "Species", "Superkingdom". Default is "Genus".}

\item{covstats}{Path to bbmap covstat output file}

\item{taxonomy}{Path to CAT taxonomy output file}
}
\value{

}
\description{
Gbtools bin plot for Shiny
}
\examples{
gbtools_plot(covstats = "/home/rstudio/data/geodescent/test_samples/bbmap/bbmap_megahit_trimmed_control_IR1-reads.fastq/covstats.txt",
taxonomy = "/home/rstudio/data/geodescent/test_samples/CAT/control_IR1-reads/taxonomy.txt")
}
