% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_ebi.R
\name{get_data_ebi}
\alias{get_data_ebi}
\title{This function downloads metadata and fastq files from ebi by MGnify accession}
\usage{
get_data_ebi(accession, outdir)
}
\arguments{
\item{accession}{MGnify accession, such as MGYS00000492}

\item{outdir}{Existing directory to store output files}

\item{ncbi_dir}{Directory where ncbi sra files are stored, edit temp dir path in ~/.ncbi/user-settings.mkfg}
}
\value{
total_md5s The md5s of all the input samples
}
\description{
This function downloads metadata and fastq files from ebi by MGnify accession
}
\examples{
get_data_ebi(accession = "MGYS00000492",
outdir = "~/outdir")
}
