% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fastq.R
\name{get_fastq}
\alias{get_fastq}
\title{This function uses the fastq_ftp col from the filereport to download the fastq files of the study.}
\usage{
get_fastq(filereport, outdir)
}
\arguments{
\item{filereport}{Data frame of filereport, only column used is fastq_ftp. This filereport can be made using get_filereport().}

\item{outdir}{Directory to store output, full path. Will be made by script (mkdir).}
}
\value{
The fastq files will be moved to the supplied output directory.
}
\description{
This function uses the fastq_ftp col from the filereport to download the fastq files of the study.
}
\examples{
get_fastq(filereport = filereport,
outdir = "/home/rstudio/data/geodescent/samples/TEST")
}
