% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filereport.R
\name{get_filereport}
\alias{get_filereport}
\title{This function downloads the filereport of a study and saves it as an rda file in extdata. The intermediate file is deleted.}
\usage{
get_filereport(url, acc)
}
\arguments{
\item{url}{Url of the filereport. The url of a filereport of a study can be gotten from the corresponding ebi metagenomic website by saving the link address of the 'TEXT' button (e.g. see https://www.ebi.ac.uk/ena/data/view/PRJEB8968)}

\item{acc}{The study accession. This accession is also mentioned in the url (e.g. "PRJEB26733")}
}
\value{
The filereport as data frame. This is also stored as rda and can be loaded using load("extdata/filename")
}
\description{
This function downloads the filereport of a study and saves it as an rda file in extdata. The intermediate file is deleted.
}
\examples{
get_filereport(url = "https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=PRJEB8968&result=read_run&fields=study_accession,sample_accession,secondary_sample_accession,experiment_accession,run_accession,tax_id,scientific_name,instrument_model,library_layout,fastq_ftp,fastq_galaxy,submitted_ftp,submitted_galaxy,sra_ftp,sra_galaxy,cram_index_ftp,cram_index_galaxy&download=txt",
acc = "PRJEB8968")
}
