% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{This functions uses the MGnify API to connect to ebi to collect the metadata for the specifies MGYS accesion}
\usage{
get_metadata(accession)
}
\arguments{
\item{accession}{Mgnify accession of project, e.g. "MGYS00000991"}
}
\value{
Data frame of metadata as collected from ebi
}
\description{
This functions uses the MGnify API to connect to ebi to collect the metadata for the specifies MGYS accesion
}
\examples{
get_metadata(accession = "MGYS00000991")
}
