% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grinder.R
\name{grinder}
\alias{grinder}
\title{Runs the biogrinder program to create artificial fastq reads from input genomes.}
\usage{
grinder(
  reference_file,
  abundance_file = "",
  fastq_output = "1",
  qual_levels = "35 10",
  mutation_dist = "poly4 3e-3 3.3e-8",
  outname,
  outdir,
  coverage = "1"
)
}
\arguments{
\item{reference_file}{Fasta file of genomes, path}

\item{abundance_file}{File of bundances of each genome for the reads, per line of file "name 22" name and perc.}

\item{fastq_output}{Default is 1, (true)}

\item{qual_levels}{Two quality values will be given: good / bad, default: "35 10"}

\item{mutation_dist}{Default is for Illumina sequences}

\item{outname}{Prefix name of output files}

\item{outdir}{Directory to store output files in}

\item{coverage}{Library coverage}
}
\value{

}
\description{
Runs the biogrinder program to create artificial fastq reads from input genomes.
}
\examples{
grinder(reference_file = "/home/$USER/scpackage/inst/extdata/GCA_002277835.1_ASM227783v1_genomic.fna",
abundance_file = "/home/$USER/abundance.txt",
outname = "test_grinder",
outdir = "grinder")
}
