% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeblastdb.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Run makeblastdb and create output folder in data}
\usage{
makeblastdb(input, outname, outdir, dbtype)
}
\arguments{
\item{input}{String, path to input fasta file}

\item{dbtype}{String, name of database type "nucl" / "prot"}

\item{out}{String, name of output files}
}
\value{

}
\description{
Run makeblastdb and create output folder in data
}
\examples{
makeblastdb(input = "/home/rstudio/data/test_db_tara",
outname = "tara_db",
outdir = "/home/rstudio/data/",
dbtype = "nucl")
}
