% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bbmap.R
\name{run_bbmap}
\alias{run_bbmap}
\title{Function to run bbmap program from r}
\usage{
run_bbmap(mode = "PE", read1, read2 = "", ref, outdir)
}
\arguments{
\item{read1}{first read fastq file path}

\item{read2}{second read fastq file path}

\item{ref}{contig fasta file}

\item{outdir}{directory to store output}
}
\value{
Creates output at specififed dir
}
\description{
Function to run bbmap program from r
}
\examples{
run_bbmap(read1 = "/home/$USER/test_trimming/ERR1424899_1P",
read2 = "/home/$USER/test_trimming/ERR1424899_2P",
ref = "/home/$USER/renamed_contigs.fa",
outdir = "/home/$USER/bbmap_test")
}
