% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CAT.R
\name{run_CAT}
\alias{run_CAT}
\title{This function runs CAT on input contigs and creates the output dir at specified location. Database download included.}
\usage{
run_CAT(samples_dir, contigpath, outname)
}
\arguments{
\item{samples_dir}{Absolute path to samples directory}

\item{contigpath}{Absolute path to the contigs fasta file}

\item{outname}{Basename for folder with output files}
}
\value{

}
\description{
This function runs CAT on input contigs and creates the output dir at specified location. Database download included.
}
