% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Function to run fastqc program on the supplied file}
\usage{
run_fastqc(filepath, outdir)
}
\arguments{
\item{filepath}{Path to fastq file}
}
\value{

}
\description{
Function to run fastqc program on the supplied file
}
\examples{
run_fastqc("~/example.fastq")
}
