% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_megahit.R
\name{run_megahit}
\alias{run_megahit}
\title{This script is a R wrapper for the bash script megahit.sh to run the megahit assembly tool}
\usage{
run_megahit(
  PE = T,
  nohup = "",
  script_path,
  r1,
  r2 = "",
  outdir,
  outname = "megahit"
)
}
\arguments{
\item{PE}{TRUE/FALSE. Wether paired end or single end library is supplied. Default = T}

\item{nohup}{can be set to "nohup" to run assembly in background. Default = ""}

\item{script_path}{Path to megahit.sh, e.g. "/home/$USER/scpackage/inst/megahit.sh"}

\item{r1}{Path to read file one, paired end or single end}

\item{r2}{Path to paired end read}

\item{outdir}{Existing directory to create output directory in}

\item{outname}{Output directory name, will be created by program}
}
\value{
A directory with output files will be created at supplied path
}
\description{
This script is a R wrapper for the bash script megahit.sh to run the megahit assembly tool
}
\examples{
run_megahit(script_path = "/home/$USER/scpackage/inst/megahit.sh",
r1 = "/home/$USER/data/geodescent/samples/MGYS00000991/ERR1424899_1.fastq.gz",
r2 = "/home/$USER/data/geodescent/samples/MGYS00000991/ERR1424899_2.fastq.gz",
outdir = "/home/$USER",
outname = "megahit_testing")
}
