% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metabat2.R
\name{run_metabat2}
\alias{run_metabat2}
\title{This function runs the metabat2 program script}
\usage{
run_metabat2(metabat_script, s_bam, fasta)
}
\arguments{
\item{metabat_script}{Path to mwetabat2 shell script}

\item{s_bam}{Sorted bam file path}

\item{fasta}{Contigs path}
}
\value{
Files created by metabat2
}
\description{
This function runs the metabat2 program script
}
