% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prokka.R
\name{run_prokka}
\alias{run_prokka}
\title{Script to activate and run prokka on metagenome}
\usage{
run_prokka(prokka_script, contigpath, outdir, prefix)
}
\arguments{
\item{prokka_script}{Path to prokka.sh, "inst/prokka.sh"}

\item{contigpath}{Path to contigs file}

\item{outdir}{Output directory}

\item{prefix}{Name for output files}
}
\value{

}
\description{
Script to activate and run prokka on metagenome
}
