% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_single.R
\name{run_sc_pipeline}
\alias{run_sc_pipeline}
\title{Shotgun metagenomic analysis and marker gene detection pipeline.}
\usage{
run_sc_pipeline(sample_path, outdir, marker_genes_fasta)
}
\arguments{
\item{marker_genes_fasta}{Path to fasta file of input marker protein(s).}

\item{samples_dir}{Path to directory with fastq files of input samples}
}
\value{

}
\description{
Shotgun metagenomic analysis and marker gene detection pipeline.
}
