% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trimmomatic.R
\name{run_trimmomatic}
\alias{run_trimmomatic}
\title{This function uses trimmomatic to trim input fastq files and outputs them to the specified base dir. Please supply input as character type.}
\usage{
run_trimmomatic(
  mode = "SE",
  f1,
  f2 = "",
  prefix = "trimmed",
  phred = "33",
  minlen = "90",
  window = "4:15",
  logfile = "trimmomatic_log.txt"
)
}
\arguments{
\item{mode}{Single end ("SE") or paired end ("PE") fastq file(s) input, default = "SE"}

\item{f1}{Absolute path to fastq read one, supply single end file / direction one PE here}

\item{f2}{Absolute path to fastq read two, only use for second paired end file not for single end file, default = "". (single end)}

\item{prefix}{Name of prefix for output files. Default = "trimmed"}

\item{phred}{Phred scale, e.g. 33 / 64. Default = "33"}

\item{minlen}{Minimum length of reads to keep, default = "90"}

\item{window}{Sliding window for quality trimming to use, default = "4:15"}

\item{outdir}{Absolute path directory to store output files, default = "./trimmomatic". Directory will be made my script.}
}
\value{
Creates trimming output files in specified directory
}
\description{
This function uses trimmomatic to trim input fastq files and outputs them to the specified base dir. Please supply input as character type.
}
